Rearrange build order to build modules first.
Kernel modules are now built before they are copied into the initramfs
and then after that, the initramfs is archived and the kernel image is
built.

When building with noconfig, do try to run make ${cfg} since no
configuration was specified and it will just trigger a regular build
prematurely.